refactor: cache location
