--- packages/app/lib/src/app/location.dart
@@ -1,14 +1,0 @@
-import 'package:geolocator/geolocator.dart';
-
-/// Determine the current position of the device.
-///
-/// When the location services are not enabled or permissions
-/// are denied the `Future` will return an error.
-
-abstract interface class LocationRepository {
-  /// Determine the current position of the device.
-  ///
-  /// When the location services are not enabled or permissions
-  /// are denied the `Future` will return an error.
-  Future<Position> determinePosition();
-}
