+++ packages/app/lib/src/features/home/application/feed_service.dart
@@ -1,7 +1,6 @@
 /// This library provides a service to stream posts in DB to the UI.
 library;
 
-import 'package:geolocator/geolocator.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 import '../../auth/application/auth_service.dart';
@@ -9,6 +8,7 @@ import '../data/post_repository.dart';
 import '../domain/feed_entity.dart';
 import '../domain/feed_model.dart';
 import '../domain/post_entity.dart';
+import 'location_service.dart';
 
 part 'feed_service.g.dart';
 
@@ -33,15 +33,12 @@ base class FeedService extends _$FeedService {
     if (cachedPost != null) return cachedPost;
 
     // Get user's location.
-    // TODO(lishaduck): Should make a location service so we cache determine location
-    final location = await determinePosition();
-    final lat = location.latitude;
-    final lng = location.longitude;
+    final location = await ref.watch(locationServiceProvider.future);
 
     final posts = await postRepo.readPosts(
       state.cursorPos,
-      lat,
-      lng,
+      location.latitude,
+      location.longitude,
     );
 
     if (posts.isEmpty) return null;
@@ -71,42 +68,4 @@ FutureOr<PostEntity?> singlePost(
   final post = await feedNotifier.fetch(postIndex);
 
   return post;
-}
-
-// TODO(lishaduck): Move this to the location repository.
-
-/// Uses geolocator plugin to request location permission and return a position
-Future<Position> determinePosition() async {
-  // Test if location services are enabled.
-  final serviceEnabled = await Geolocator.isLocationServiceEnabled();
-  if (!serviceEnabled) {
-    // Location services are not enabled don't continue
-    // accessing the position and request users of the
-    // App to enable the location services.
-    throw Exception('Location services are disabled.');
-  }
-
-  var permission = await Geolocator.checkPermission();
-  if (permission == LocationPermission.denied) {
-    permission = await Geolocator.requestPermission();
-    if (permission == LocationPermission.denied) {
-      // Permissions are denied, next time you could try
-      // requesting permissions again (this is also where
-      // Android's shouldShowRequestPermissionRationale
-      // returned true. According to Android guidelines
-      // your App should show an explanatory UI now.
-      throw Exception('Location permissions are denied');
-    }
-  }
-
-  if (permission == LocationPermission.deniedForever) {
-    // Permissions are denied forever, handle appropriately.
-    throw Exception(
-      'Location permissions are permanently denied, we cannot request permissions.',
-    );
-  }
-
-  // When we reach here, permissions are granted and we can
-  // continue accessing the position of the device.
-  return Geolocator.getCurrentPosition();
 }
