+++ packages/app/lib/src/features/home/data/location_repository.dart
@@ -1,0 +1,61 @@
+import 'package:geolocator/geolocator.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+part 'location_repository.g.dart';
+
+/// Provide access to the user's location.
+abstract interface class LocationRepository {
+  /// Determine the current position of the device.
+  ///
+  /// When the location services are not enabled or permissions
+  /// are denied the `Future` will return an error.
+  Future<Position> determinePosition();
+}
+
+final class _GeolocatorLocationRepository implements LocationRepository {
+  const _GeolocatorLocationRepository();
+
+  @override
+
+  /// Uses geolocator plugin to request location permission and return a position
+  Future<Position> determinePosition() async {
+    // Test if location services are enabled.
+    final serviceEnabled = await Geolocator.isLocationServiceEnabled();
+    if (!serviceEnabled) {
+      // Location services are not enabled don't continue
+      // accessing the position and request users of the
+      // App to enable the location services.
+      throw Exception('Location services are disabled.');
+    }
+
+    var permission = await Geolocator.checkPermission();
+    if (permission == LocationPermission.denied) {
+      permission = await Geolocator.requestPermission();
+      if (permission == LocationPermission.denied) {
+        // Permissions are denied, next time you could try
+        // requesting permissions again (this is also where
+        // Android's shouldShowRequestPermissionRationale
+        // returned true. According to Android guidelines
+        // your App should show an explanatory UI now.
+        throw Exception('Location permissions are denied');
+      }
+    }
+
+    if (permission == LocationPermission.deniedForever) {
+      // Permissions are denied forever, handle appropriately.
+      throw Exception(
+        'Location permissions are permanently denied, we cannot request permissions.',
+      );
+    }
+
+    // When we reach here, permissions are granted and we can
+    // continue accessing the position of the device.
+    return Geolocator.getCurrentPosition();
+  }
+}
+
+/// Get the user's location.
+@riverpod
+LocationRepository locationRepository(LocationRepositoryRef ref) {
+  return const _GeolocatorLocationRepository();
+}
